Remove all MAX7456 specific code, replace it with generic interfaces
- Rename max7456_symbols.h to osd_symbols.h. We'll expect all supported
OSD drivers to use the same font style.
- Introduce drivers/osd.h, which defines the supported video systems
and the format for the OSD characters that all drivers must support.
- Remove drivers/vcd.h, since having a file for a single enum is
kind of annoying, move this enum to the new drivers/osd.h file.
- Add osdGetDisplayPort() for retriving the display port used by
the OSD.
- Add a new display function named writeFontCharacter() which takes
a character address and its data.
- Use the aforementioned 2 functions to implement character font
uploading in fc_msp.c, so it's not MAX7456-specific anymore.